requirements.txt
absl-py==1.0.0
astor==0.8.1
cached-property==1.5.2
cycler==0.11.0
fonttools==4.29.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.44.0
h5py==2.10.0
importlib-metadata==4.11.1
imutils==0.5.4
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
mock==4.0.3
numpy==1.21.5
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
protobuf==3.19.4
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
scipy==1.7.3
six==1.16.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
typing_extensions==4.1.1
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
